// frontend-admin/src/api/generated/model/replayBatchPaymentItemDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */

export interface ReplayBatchPaymentItemDto {
  createdAtUtc?: string;
  /** @nullable */
  offlineTransactionId?: string | null;
  paymentId?: string;
  /** @nullable */
  receiptId?: string | null;
  /** @nullable */
  receiptNumber?: string | null;
  totalAmount?: number;
}